requirements.txt
asn1crypto==0.24.0
awsebcli==3.14.6
azure-common==1.1.15
azure-nspkg==2.0.0
azure-storage==0.36.0
Babel==2.6.0
bcrypt==3.1.4
blinker==1.4
boto3==1.8.9
botocore==1.12.0
cement==2.8.2
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
cryptography==2.3.1
docutils==0.14
Flask==1.0.2
Flask-Admin==1.5.2
Flask-BabelEx==0.9.3
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.16.0
gunicorn==19.9.0
idna==2.7
ijson==2.3
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
passlib==1.7.1
pathspec==0.5.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
pycryptodomex==3.6.6
PyJWT==1.6.4
pyOpenSSL==18.0.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
s3transfer==0.1.13
semantic-version==2.5.0
six==1.11.0
snowflake-connector-python==1.6.9
snowflake-sqlalchemy==1.1.2
speaklater==1.3
SQLAlchemy==1.2.12
termcolor==1.1.0
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.2.1
